refactor(Endpoints): add new hooks and components for endpoint management
