+++ client/src/components/Endpoints/ConvoIcon.tsx
@@ -1,8 +1,8 @@
 import React, { useMemo } from 'react';
 import type * as t from 'librechat-data-provider';
 import { getEndpointField, getIconKey, getEntity, getIconEndpoint } from '~/utils';
-import { icons } from '~/components/Chat/Menus/Endpoints/Icons';
 import ConvoIconURL from '~/components/Endpoints/ConvoIconURL';
+import { icons } from '~/hooks/Endpoint/Icons';
 
 export default function ConvoIcon({
   conversation,
