+++ client/src/components/Endpoints/MinimalIcon.tsx
@@ -11,7 +11,7 @@ import {
   BedrockIcon,
   Sparkles,
 } from '~/components/svg';
-import UnknownIcon from '~/components/Chat/Menus/Endpoints/UnknownIcon';
+import UnknownIcon from '~/hooks/Endpoint/UnknownIcon';
 import { IconProps } from '~/common';
 import { cn } from '~/utils';
