+++ client/src/components/SidePanel/Agents/AgentConfig.tsx
@@ -6,9 +6,9 @@ import type { TPlugin } from 'librechat-data-provider';
 import type { AgentForm, AgentPanelProps, IconComponentTypes } from '~/common';
 import { cn, defaultTextProps, removeFocusOutlines, getEndpointField, getIconKey } from '~/utils';
 import { useToastContext, useFileMapContext } from '~/Providers';
-import { icons } from '~/components/Chat/Menus/Endpoints/Icons';
 import Action from '~/components/SidePanel/Builder/Action';
 import { ToolSelectDialog } from '~/components/Tools';
+import { icons } from '~/hooks/Endpoint/Icons';
 import { processAgentOption } from '~/utils';
 import AgentAvatar from './AgentAvatar';
 import FileContext from './FileContext';
