+++ client/src/hooks/index.ts
@@ -17,6 +17,7 @@ export * from './AuthContext';
 export * from './ThemeContext';
 export * from './ScreenshotContext';
 export * from './ApiErrorBoundaryContext';
+export * from './Endpoint';
 
 export type { TranslationKeys } from './useLocalize';
